♻️ Refactor: 추가 정보 입력 페이지 변경
